Add value version numbers (issue #2).
This commit introduces the relevant methods and fields and makes the
modifications needed in order to deal with version numbers. However, it does not
actually enable the version number support because the version number sources
are not set yet, which has the consequence of falling back to the old behavior
that was used before version numbers were introduced.