Added the version number source to the PV manager (issue #2).
The PV manager uses a global version number source and passes it to its process
variables. Tests that modify the process variable's value by getting through the
reference returned by get() are broken by this change. However, this is not
fixed here because the version number mechanism is most likely going to be
redesigned anyway.