Add modified methods to process array (issue #2).
With the modified method, code modifying the array through a non-const reference
to the value can mark it as modified so that a new version number is assigned
and it is synchronized correctly.

In addition to that, a modification was made to the PVManager test so that a
scalar is modified in order to actually trigger a time-stamp update.